Fixed couple errors & Updated README

+++ back/index.js
@@ -6,9 +6,16 @@ const TodoModel = require('./Models/Todo')
 const app = express()
 app.use(cors())
 app.use(express.json())
+
+const port = 3000
 const uri = "mongodb+srv://test:test@todo-react.wppa9ep.mongodb.net/todo-react?retryWrites=true&w=majority&appName=todo-react";
 const clientOptions = { serverApi: { version: '1', strict: true, deprecationErrors: true } };
-mongoose.connect(uri, clientOptions);
+try {
+	mongoose.connect(uri, clientOptions);
+	console.log("Database connected");
+} catch {
+	console.log("Database failed");
+}
 
 app.get("/get", (req, res) => {
   TodoModel.find()
@@ -17,22 +24,34 @@ app.get("/get", (req, res) => {
 })
 app.post("/add", (req, res) => {
 	const task = req.body.task;
-	TodoModel.create({
-		task: task
-	})
-	.then(result => res.json(result))
-	.catch(err => res.json(err))
+	if (task) {
+		TodoModel.create({
+			task: task
+		})
+		.then(result => res.json(result))
+		.catch(err => res.json(err))
+	} else {
+		res.json({err: "No name"})
+	}
 })
 app.put("/update/:id", (req, res) => {
 	const {id} = req.params;
-	TodoModel.findByIdAndUpdate({_id: id}, req.body.todo)
-	.then(result => res.json(result))
-	.catch(err => res.json(err))
+	if (id) {
+		TodoModel.findByIdAndUpdate({_id: id}, req.body.todo)
+		.then(result => res.json(result))
+		.catch(err => res.json(err))
+	} else {
+		res.json({err: "No ID"})
+	}
 })
 app.delete("/delete/:id", (req, res) => {
 	const {id} = req.params;
-	TodoModel.deleteOne({_id: id})
-	.then(result => res.json(result))
-	.catch(err => res.json(err))
+	if (id) {
+		TodoModel.deleteOne({_id: id})
+		.then(result => res.json(result))
+		.catch(err => res.json(err))
+	} else {
+		res.json({err: "No ID"})
+	}
 })
-app.listen(3000, () => console.log("Server is running"))
+app.listen(port, () => console.log("Server is running. Port: " + port))
